+++ python/settings.cpp
@@ -60,7 +60,8 @@ void init_settings(py::module_ &m) {
         .def_readwrite("domFiles", &WorldOptions::domFiles)
         .def_readwrite("logLevel", &WorldOptions::logLevel)
         .def_readwrite("rasterUsageThreshold", &WorldOptions::rasterUsageThreshold)
-        .def_readwrite("minRes", &WorldOptions::minRes);
+        .def_readwrite("minRes", &WorldOptions::minRes)
+        .def_readwrite("maxRes", &WorldOptions::maxRes);
 
     /* RAYTRACER OPTIONS */
     py::class_<RayTracerOptions>(m, "RayTracerOptions")