+++ src/atlas.cpp
@@ -5,6 +5,7 @@
 #include <filesystem>
 
 #include "opencv2/imgcodecs.hpp"
+#include "opencv2/opencv.hpp"
 
 // Utility function to ensure images have the proper bits
 void checkImageBits(int type) {
@@ -102,6 +103,18 @@ cv::Mat RayTracer::createImageOptical(int type) {
 
     const std::vector<RenderedPixel>* pixels = renderer.getRenderedPixels();
 
+    /* For the DOM I should use the minimum pixel resolution, so that I ensure the
+     * lighting is consistent as much as possible across the image. So we iterate 
+     * across all pixels and retrieve the minimum ray resolution. */
+    auto it = std::min_element(
+        pixels->begin(), pixels->end(),
+        [](const RenderedPixel& a, const RenderedPixel& b) {
+            return a.pixResolution() < b.pixResolution();
+        }
+    );
+
+    double minRayRes = it->pixResolution();
+
     // Create a grayscale image (8-bit single-channel)
     cv::Mat image(cam->height(), cam->width(), type, cv::Scalar(0));
 
@@ -126,7 +139,7 @@ cv::Mat RayTracer::createImageOptical(int type) {
                 s[1] = rad2deg(p.data[k].s[2]); 
                 
                 // Sample the DOM at that location
-                c += world.sampleDOM(s, p.pixResolution());
+                c += world.sampleDOM(s, minRayRes);
             }
         }
 
@@ -279,6 +292,58 @@ cv::Mat RayTracer::createDepthMap(int type) {
     }
 
     return image; 
+
+}
+
+cv::Mat RayTracer::createLIDARMap() {
+
+    // Check camera pointer 
+    checkCamPointer();
+    // Check rendering status
+    checkRenderStatus();
+
+    const std::vector<RenderedPixel>* pixels = renderer.getRenderedPixels();
+
+    // Create two single channel floating point images
+    cv::Mat lidar(cam->height(), cam->width(), CV_64F, cv::Scalar(0));
+    cv::Mat elev(cam->height(), cam->width(), CV_64F, cv::Scalar(0));
+
+    ui32_t u, v; 
+    double depth, elevation;
+    int i;  
+
+    for (auto& p: *pixels) {
+
+        // Retrieve the pixel coordinates
+        cam->getPixelCoordinates(p.id, u, v); 
+
+        // Compute the average pixel depth 
+        depth = 0.0; elevation = 0.0; i = 0;
+        for (size_t k = 0; k < p.nSamples; k++) {
+            if (p.data[k].t != inf) {
+                depth += p.data[k].t;           // Update the depth distance
+                elevation += p.data[k].s[0];    // Update the elevation values
+                i += 1;
+            }
+        }
+        
+        if (i > 0) {
+            // Average the distance through all the valid samples
+            lidar.at<double>(v, u) = depth / i;
+            elev.at<double>(v, u) = elevation / i;
+
+        } else {
+
+            // Assign nan in that place 
+            lidar.at<double>(v, u) = std::numeric_limits<double>::quiet_NaN(); 
+            elev.at<double>(v, u) = std::numeric_limits<double>::quiet_NaN();
+        }
+    }
+
+    // Merge the two channels into a two-channel image
+    cv::Mat image;
+    cv::merge(std::vector<cv::Mat>{lidar, elev}, image);
+    return image;
 
 }