+++ src/raster.cpp
@@ -99,6 +99,7 @@ void RasterBand::loadData() {
     );
 
     if (err != CE_None) {
+        unloadData();
         throw std::runtime_error("failed to retrieve raster band data");
     } 
 
@@ -106,8 +107,13 @@ void RasterBand::loadData() {
 }
 
 void RasterBand::unloadData() {
+    
+    /* Frees the internal GDAL cache*/
+    pBand->FlushCache();
+
     nLoadedElements = 0; 
     data.reset();
+    
 }
 
 double RasterBand::getData(ui32_t i) const {
@@ -566,6 +572,12 @@ double RasterManager::getData(const point2& s, double res, ui32_t tid) {
     // Set the initial impact distance to -inf (i.e., no data available)
     double x = -inf;
 
+    // Ensure at least one raster is loaded, otherwise return infinity.
+    if (_nRasters == 0) {
+        std::clog << "No rasters have been loaded in the container." << std::endl;
+        return x;
+    }
+    
     // Find the last container with a resolution lower than the prescribed one.
     size_t cIdx = findLast(_resolutions, res);