+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # All Releases 
 
 ## v0.6.0
+- Added `createLIDARMap` to export pixel depths and elevation values as an image in physical units (meters).
 - Added `ScreenGrid` to allow multiple ray resolutions over the same image.
 - Updated `RenderingOptions` with `gridWidth` and `gridHeight` parameters.
 - Added `minRes` parameter to `WorldOptions` to control the minimum ray resolution.
 - Fixed GSD computation when below the average planet radius.
+- Exposed world class via raytracer through `getWorld`.
+- Fixed segmentation error with `getData` when no rasters are loaded.
+- Updated raster unloading.
 
 ## v0.5 
 - Renamed `LunarRayTracer` to `RayTracer`.